grill/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
branca==0.4.2
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
distlib==0.3.3
Django==3.2.9
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.13.0
django-ipware==4.0.0
djangorestframework==3.12.4
filelock==3.4.0
folium==0.12.1.post1
future==0.18.2
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
telegram==0.0.1
telegram-send==0.25
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.10.0
